fix: remove FilterContext move logic to AnnouncementsContext
