+++ src/context/AnnouncementsContext/index.tsx
@@ -1,17 +1,18 @@
-import { createContext } from "react";
-import {
-  IAnnouncementsContext,
-  IAnnouncementsForm,
-  IAnnouncementsProviderProps,
-} from "./types";
+import { createContext, useState } from "react";
 import { api } from "../../services/api";
+import { IAnnouncementsContext, IAnnouncementsForm, IAnnouncementsProviderProps, IFilterData, IFilterResponse } from "../../interfaces/announcementsContext.types";
+import { AxiosResponse } from "axios";
 
 export const AnnouncementsContext = createContext<IAnnouncementsContext>(
   {} as IAnnouncementsContext
 );
 
 const AnnouncementsProvider = ({ children }: IAnnouncementsProviderProps) => {
-  
+  const [Announcements, SetAnnouncements] = useState<IFilterResponse[]>([]);
+  const [allAnnouncementsForFilter, SetAllAnnouncementsForFilter] = useState<
+    IFilterResponse[]
+  >([]);
+
   const createAnnouncement = async (
     dataAnnouncement: IAnnouncementsForm
   ): Promise<void> => {
@@ -37,10 +38,66 @@ const AnnouncementsProvider = ({ children }: IAnnouncementsProviderProps) => {
     }
   };
 
+  const getAnnouncements = async (data: IFilterData) => {
+    const {
+      brand,
+      color,
+      fuel,
+      maxKm,
+      model,
+      maxPrice,
+      minKm,
+      minPrice,
+      year,
+    } = data;
+
+    try {
+      const response: AxiosResponse<IFilterResponse[]> = await api.get(
+        `announcements?brand=${brand}&model=${model}&color=${color}&year=${year}&minKm=${minKm}&max
+            Km=${maxKm}&minPrice=${minPrice}&maxPrice=${maxPrice}&fuel=${fuel}`
+      );
+      console.log(response.data);
+
+      SetAnnouncements(response.data);
+    } catch (error) {
+      console.error(error);
+    }
+  };
+
+  const getAllAnnouncementsForFilter = async (data: IFilterData) => {
+    const {
+      brand,
+      color,
+      fuel,
+      maxKm,
+      model,
+      maxPrice,
+      minKm,
+      minPrice,
+      year,
+    } = data;
+
+    try {
+      const response: AxiosResponse<IFilterResponse[]> = await api.get(
+        `announcements?brand=${brand}&model=${model}&color=${color}&year=${year}&minKm=${minKm}&max
+            Km=${maxKm}&minPrice=${minPrice}&maxPrice=${maxPrice}&fuel=${fuel}`
+      );
+      console.log(response.data);
+
+      SetAllAnnouncementsForFilter(response.data);
+    } catch (error) {
+      console.error(error);
+    }
+  };
+
   return (
     <AnnouncementsContext.Provider
       value={{
         createAnnouncement,
+        getAnnouncements,
+        getAllAnnouncementsForFilter,
+        allAnnouncementsForFilter,
+        Announcements
       }}
     >
       {children}
