+++ src/pages/home/index.tsx
@@ -1,47 +1,35 @@
-import { useContext, useEffect, useState } from "react";
-import { FilterContext } from "../../context/FilterContext";
+import { useEffect } from "react";
 import { ProductCard } from "../../components/ProductCard";
 import { Filter } from "../../components/Filter";
-import AnnouncementForm from "../../components/Forms/AnnouncementForm";
 import useAnnouncements from "../../hooks/useAnnouncements";
 
 export const Home = () => {
-    const { createAnnouncement } = useAnnouncements();
-    const [openAnnouncementModal, setOpenAnnouncementModal] = useState(false);
+  const { Announcements, getAnnouncements } = useAnnouncements();
 
-   const { Announcements, getAnnouncements } = useContext(FilterContext);
-   
-      const dataTeste = {
-      brand: "",
-      model: "",
-      year: "",
-      fuel: "",
-      color: "",
-      minPrice: "",
-      maxPrice: "",
-      minKm:"",
-      maxKm:"",
-   }
+  const dataTeste = {
+    brand: "",
+    model: "",
+    year: "",
+    fuel: "",
+    color: "",
+    minPrice: "",
+    maxPrice: "",
+    minKm: "",
+    maxKm: "",
+  };
 
-   useEffect(() => {
-      getAnnouncements(dataTeste);
-   }, []);
+  useEffect(() => {
+    getAnnouncements(dataTeste);
+  }, []);
 
-   return (
-      <main>
-         <Filter/>
-         <ul>
-            {Announcements.map((announcement) => (
-               <ProductCard key={announcement.id} announcement={announcement} />
-            ))}
-         </ul>
-         <AnnouncementForm
-        isCreateForm={true}
-        open={openAnnouncementModal}
-        setOpen={setOpenAnnouncementModal}
-        submitFunction={createAnnouncement}
-      />
-      </main>
-)
-}
-
+  return (
+    <main>
+      <Filter />
+      <ul>
+        {Announcements.map((announcement) => (
+          <ProductCard key={announcement.id} announcement={announcement} />
+        ))}
+      </ul>
+    </main>
+  );
+};
