--- src/services/api/index.ts
@@ -1,6 +1,0 @@
-import axios from "axios";
-
-export const api = axios.create({
-   baseURL: 'http://localhost:3000/',
-   timeout: 10000,
-});
